requirement.txt
google-generativeai
python-dotenv
openai-whisper
transformers
torch
Pillow
pandas
scikit-learn
h5py
opencv-python
openai
transformers
youtube_transcript_api
fuzzywuzzy
pytube
yt-dlp